% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{asympt_draw}
\alias{asympt_draw}
\title{Draw from the asymptotic distribution of the estimate}
\usage{
asympt_draw(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=asympt_draw.SharedModel]{asympt_draw.SharedModel()}} for more information.
}
