% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{asympt_draw.SharedModel}
\alias{asympt_draw.SharedModel}
\title{Draw from the asymptotic distribution of the estimate}
\usage{
\method{asympt_draw}{SharedModel}(obj, n = 100, enforce_posit = TRUE)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}

\item{n}{Sample size.}

\item{enforce_posit}{\code{TRUE} to enforce positivity (default).}
}
\value{
n times number of parameters matrix.
}
\description{
Draw \code{n} points from the asymptotic distribution of the estimated parameter
of the model. It takes into account positivity constraints.
}
\details{
If \code{enforce_posit=TRUE}, the function ensures positivity of some parameters
(e.g., Weibull scale), even if this was not enforced during optimization.
This essentially draws from the asymptotic distribution of the parameters,
conditonal of these being positive.
}
