% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{asymptvar}
\alias{asymptvar}
\title{Asymptotic covariance matrix}
\usage{
asymptvar(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=asymptvar.SharedModel]{asymptvar.SharedModel()}} for more information.
}
