% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{asymptvar.SharedModel}
\alias{asymptvar.SharedModel}
\title{Asymptotic covariance matrix}
\usage{
\method{asymptvar}{SharedModel}(obj, BHHH = TRUE)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}

\item{BHHH}{Whether to approximate the Information Matrix by the sum of
squared scores. Values = \code{TRUE} (default) or \code{FALSE}.}
}
\value{
numeric matrix with asymptotic covariances
}
\description{
Computes the asymptotic covariance matrix of the fitted parameters.
}
\details{
The asymptotic covariance matrix is the inverse of the Information Matrix.

\code{BHHH = TRUE} uses the sum of squared scores to estimate the Information Matrix.
\code{BHHH = FALSE} uses the Hessian returned by \code{\link[stats:nlm]{stats::nlm()}} to estimate the
Information Matrix. Note that the \code{BHHH = FALSE} method requires small
scores to obtain a correct approximation (due to transformations that ensure
positivity of certain parameters).
}
