% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{base_hazard}
\alias{base_hazard}
\title{Baseline hazard functions}
\usage{
base_hazard(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=base_hazard.SharedModel]{base_hazard.SharedModel()}} for more information.
}
