% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{base_survival}
\alias{base_survival}
\title{Baseline survival functions}
\usage{
base_survival(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=base_survival.SharedModel]{base_survival.SharedModel()}} for more information.
}
