% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{base_survival.SharedModel}
\alias{base_survival.SharedModel}
\title{Baseline survival functions}
\usage{
\method{base_survival}{SharedModel}(obj, t, process, gradient = FALSE)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}

\item{t}{time to evaluate the survival function}

\item{process}{Process for which the survival function is computed. Either
\code{terminal} (\code{t}) or \code{recurrent} (\code{r}).}

\item{gradient}{\code{TRUE} to compute the gradient (default is \code{FALSE}).}
}
\value{
numeric value of the survival function at \code{t}. If \code{gradient=TRUE},
the returned value has an attribute \code{"gradient"} with its derivative.
}
\description{
Get the baseline survival functions at evaluated at \code{t}.
}
