% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{BHHH_hessian}
\alias{BHHH_hessian}
\title{Berndt–Hall–Hall–Hausman Hessian}
\usage{
BHHH_hessian(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=BHHH_hessian.SharedModel]{BHHH_hessian.SharedModel()}} for more information.
}
