% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{BHHH_hessian.SharedModel}
\alias{BHHH_hessian.SharedModel}
\title{Berndt–Hall–Hall–Hausman Hessian}
\usage{
\method{BHHH_hessian}{SharedModel}(obj, S = NULL)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}

\item{S}{matrix with scores (individuals in rows). Defaul is \code{NULL}, so that
the function computes the matrix.}
}
\value{
Estimate of the Asymptotic Information matrix
}
\description{
Estimates the Asymptotic Information matrix by aggregating the squared
scores.
}
