% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{check_intsample}
\alias{check_intsample}
\title{Check and compute numerical integral sample}
\usage{
check_intsample(obj)
}
\description{
This is an S3 generic. See \code{\link[=check_intsample.SharedModel]{check_intsample.SharedModel()}} for more information.
}
