% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{check_intsample.SharedModel}
\alias{check_intsample.SharedModel}
\title{Check whether there is sample for numerical integration}
\usage{
\method{check_intsample}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
\code{\link[=SharedModel]{SharedModel()}} object with a stored numerical integration sample.
}
\description{
Check whether there is Monte Carlo sample is stored. If not, it computes it.
}
