% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{dist_relevance_test}
\alias{dist_relevance_test}
\title{Test relevance of recurrent event temporal distribution}
\usage{
dist_relevance_test(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=dist_relevance_test.SharedModel]{dist_relevance_test.SharedModel()}} for more information.
}
