% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{dist_relevance_test.SharedModel}
\alias{dist_relevance_test.SharedModel}
\title{Test relevance of recurrent event temporal distribution}
\usage{
\method{dist_relevance_test}{SharedModel}(obj, verbose = FALSE)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}

\item{verbose}{}
}
\description{
!!!!!! NEED TO DOCUMENT THIS.
}
