% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{get_rec_number}
\alias{get_rec_number}
\title{Find number of recurrent events}
\usage{
get_rec_number(obj)
}
\description{
This is an S3 generic. See \code{\link[=get_rec_number.SharedModel]{get_rec_number.SharedModel()}} for more information.
}
