% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{get_rec_number.SharedModel}
\alias{get_rec_number.SharedModel}
\title{Number of recurrent events for each individual}
\usage{
\method{get_rec_number}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
numeric vector with the number of recurrent event for each individual.
}
\description{
Number of recurrent events for each individual
}
