% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{hazard_given_history}
\alias{hazard_given_history}
\title{Hazard function given recurrent event history}
\usage{
hazard_given_history(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=hazard_given_history.SharedModel]{hazard_given_history.SharedModel()}} for more information.
}
