% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{hazard_given_history.SharedModel}
\alias{hazard_given_history.SharedModel}
\title{Hazard given history}
\usage{
\method{hazard_given_history}{SharedModel}(obj, t, Tjs = NULL, gradient = FALSE)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}

\item{t}{time to evaluate the hazard}

\item{Tjs}{numeric vector with recurrent event times prior to \code{t}.}

\item{gradient}{\code{TRUE} to compute the gradient (default is \code{FALSE}).}
}
\value{
numeric value of the hazard given history at \code{t}. If \code{gradient=TRUE},
the returned value has an attribute \code{"gradient"} with its derivative.
}
\description{
Finds the hazard of the recurrent event at \code{t}, given the history of
recurrent events prior to \code{t}.
}
