% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fit.R
\name{iloglik_intfrailty}
\alias{iloglik_intfrailty}
\title{Part of the loglikelihood that integrates w.r.t frailty (u)}
\usage{
iloglik_intfrailty(obs, model)
}
\arguments{
\item{obs}{Observation unit information. See \code{\link[=individual_loglikelihood]{individual_loglikelihood()}}.}

\item{model}{\code{\link[=SharedModel]{SharedModel()}} object with model information.}
}
\value{
numeric Integral part of the log-likelihood
}
\description{
Part of the loglikelihood that integrates w.r.t frailty (u)
}
