% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fit.R
\name{iloglik_intfrailty_gradient}
\alias{iloglik_intfrailty_gradient}
\title{Derivative of the part of the loglikelihood with the integral wrt frailty}
\usage{
iloglik_intfrailty_gradient(obs, model)
}
\arguments{
\item{obs}{Observation unit information. See \code{\link[=individual_loglikelihood]{individual_loglikelihood()}}.}

\item{model}{\code{\link[=SharedModel]{SharedModel()}} object with model information.}
}
\value{
numeric vector Derivative of the integral part of the log-likelihood
}
\description{
Derivative of the part of the loglikelihood with the integral wrt frailty
}
