% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fit.R
\name{iloglik_nofrailty}
\alias{iloglik_nofrailty}
\title{Part of the log-likelihood that does not depend on the frailty}
\usage{
iloglik_nofrailty(obs, model)
}
\arguments{
\item{obs}{Observation unit information. See \code{\link[=individual_loglikelihood]{individual_loglikelihood()}}.}

\item{model}{\code{\link[=SharedModel]{SharedModel()}} object with model information.}
}
\value{
Part of the log-likelihood that does not depend on the frailty
}
\description{
Part of the log-likelihood that does not depend on the frailty
}
