% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fit.R
\name{individual_loglikelihood}
\alias{individual_loglikelihood}
\title{Log-likelihood computation for an individual}
\usage{
individual_loglikelihood(obs, model)
}
\arguments{
\item{obs}{Observation unit information. List with components:
\itemize{
\item \verb{$ter_time} = Time and censoring indicator for terminal process.
\item \verb{$ter_covs} = Matrix with covariates in terminal process model.
\item \verb{$rec_times} =  Times and censoring indicators for recurrent process.
\item \verb{$rec_covs} = Matrix with covariates in recurrent process model.
}}

\item{model}{\code{\link[=SharedModel]{SharedModel()}} object with model information.}
}
\value{
numeric with log-likelihooh value.
}
\description{
Computes the log-likelihood for an observation unit in the sample.
}
