% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fit.R
\name{individual_score}
\alias{individual_score}
\title{Score computation for and individual}
\usage{
individual_score(obs, model)
}
\arguments{
\item{obs}{Observation unit information. See \code{\link[=individual_loglikelihood]{individual_loglikelihood()}}.}

\item{model}{\code{\link[=SharedModel]{SharedModel()}} object with model information.}
}
\value{
numeric vector Score
}
\description{
Score computation for and individual
}
