% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{logLik.SharedModel}
\alias{logLik.SharedModel}
\title{Log-likelihood of the model}
\usage{
\method{logLik}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
numeric value of the log-likelihood
}
\description{
Computes the log-likelihood of the model.
}
