% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fit.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Log-likelihood function for the whole data}
\usage{
loglikelihood(
  nu,
  OBSL,
  par_pos,
  posit_cons,
  model_funs,
  rec_timescale,
  int_mode,
  MC_N,
  GC_nodes,
  int_seq,
  is_control,
  anal.grad,
  BHHH.hessian
)
}
\arguments{
\item{nu}{Guess value of the parameter}

\item{OBSL}{List with all observations}

\item{par_pos}{Position of each specific parameter in nu}

\item{posit_cons}{Elements in nu subject to positivity constraints}

\item{model_funs}{List with model function shapes}

\item{rec_timescale}{Timescale for the recurrent events}

\item{int_mode}{How to perform the integral}

\item{MC_N}{Monte Carlo sample size}

\item{GC_nodes}{Gaussian Quadrature number of nodes}

\item{int_seq}{Additional sequence needed for integration in some modes}

\item{is_control}{Control for Importance Sampling: distribution and parameters}

\item{BHHH.hessian}{Whether to compute BHHH style hessian or not}

\item{anal.grand}{Whether to use the analytic gradient or no}
}
\value{
Log-likelihood as numeric with attribute 'gradient' and 'hessian'
}
\description{
Computes the log-likelihood for the whole data given a parameter guess nu.
Input for optimization via \code{\link[stats:nlm]{stats::nlm()}}.
}
