% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{max_timespan}
\alias{max_timespan}
\title{Largest time observation in the sample}
\usage{
max_timespan(obj)
}
\description{
This is an S3 generic. See \code{\link[=max_timespan.SharedModel]{max_timespan.SharedModel()}} for more information.
}
