% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{max_timespan.SharedModel}
\alias{max_timespan.SharedModel}
\title{Largest time observation in the sample}
\usage{
\method{max_timespan}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
numeric value of the largest time observation in the sample
}
\description{
Largest time observation in the sample
}
