% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{optim_details}
\alias{optim_details}
\title{Optimization details}
\usage{
optim_details(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=optim_details.SharedModel]{optim_details.SharedModel()}} for more information.
}
