% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{optim_details.SharedModel}
\alias{optim_details.SharedModel}
\title{Optimization details}
\usage{
\method{optim_details}{SharedModel}(obj, digits = 1)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
Silent. Prints in console.
}
\description{
Print optimization details in console.
}
