% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{param_coef}
\alias{param_coef}
\title{Coefficients of covariates}
\usage{
param_coef(obj)
}
\description{
This is an S3 generic. See \code{\link[=param_coef.SharedModel]{param_coef.SharedModel()}} for more information.
}
