% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{param_coef.SharedModel}
\alias{param_coef.SharedModel}
\title{Coefficients of covariates}
\usage{
\method{param_coef}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
List with the coefficients:
\itemize{
\item \verb{$terminal} = for the terminal event process.
\item \verb{$recurrent} = for the recurrent event process.
}
}
\description{
Coefficients of covariates
}
