% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{param_frailty}
\alias{param_frailty}
\title{Parameters relative to the frailty term}
\usage{
param_frailty(obj)
}
\description{
This is an S3 generic. See \code{\link[=param_frailty.SharedModel]{param_frailty.SharedModel()}} for more information.
}
