% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{param_frailty.SharedModel}
\alias{param_frailty.SharedModel}
\title{Parameters relative to the frailty term}
\usage{
\method{param_frailty}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
named numeric vector with the coefficients
}
\description{
Parameters relative to the frailty term
}
