% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{param_hazard}
\alias{param_hazard}
\title{Parameters of the baseline hazards}
\usage{
param_hazard(obj)
}
\description{
This is an S3 generic. See \code{\link[=param_hazard.SharedModel]{param_hazard.SharedModel()}} for more information.
}
