% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{param_hazard.SharedModel}
\alias{param_hazard.SharedModel}
\title{Parameters of the baseline hazards}
\usage{
\method{param_hazard}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
List with the parameters:
\itemize{
\item \verb{$terminal} = for the terminal event process.
\item \verb{$recurrent} = for the recurrent event process.
}
}
\description{
Parameters of the baseline hazards
}
