% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{param_tables.SharedModel}
\alias{param_tables.SharedModel}
\title{Parameter tables for printing and latex output}
\usage{
\method{param_tables}{SharedModel}(
  obj,
  cols = c("est", "est_SE", "hr", "hr_SE", "pvals"),
  col_names = NULL,
  BHHH = TRUE
)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object with fit results.}

\item{cols}{character vector with values to display. A subset of
\code{c('est', 'est_SE', 'hr', 'hr_SE', 'pvals')}.}

\item{col_names}{Names of the columns.}

\item{BHHH}{\code{TRUE} to use Score based variance for computations. \code{FALSE} to
use the variance based on the optimization routine Hessian.}
}
\value{
List with tables.
\itemize{
\item \verb{$beta_d} = Covariate coefficients for terminal event hazard.
\item \verb{$beta_r} = Covariate coefficients for recurrent event hazard.
\item \verb{$a_d} = Parameters of baseline hazard of terminal event.
\item \verb{$a_r} = Parameters of baseline hazard of recurrent event.
\item \verb{$frail} = Parameters of frailty pdf.
}
}
\description{
Parameter tables for printing and latex output
}
