% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fit.R
\name{posit_constraint}
\alias{posit_constraint}
\title{Impose positivity constraint}
\usage{
posit_constraint(par, impose_ind, type = "direct")
}
\arguments{
\item{par}{Parameter of the model, see \code{\link[=SharedModel]{SharedModel()}}.}

\item{impose_ind}{Indexes to impose the constraint, see \code{\link[=SharedModel]{SharedModel()}}.}

\item{type}{Either \code{"direct"} (positive-valued function: \code{x^2}), \code{"reverse"}
(inverse of the function), \code{"gradient"} (gradient of the transformation), or
\code{"diag_hessian"} (diagonal of the Hessian of the transformation)}
}
\value{
numeric vector Depending on \code{type}.
}
\description{
Function to impose and reverse the square transformation
for positivity constraint.
}
