% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_frailty_pred.R
\name{predict.SharedModel}
\alias{predict.SharedModel}
\title{Predict risk of terminal event}
\usage{
\method{predict}{SharedModel}(
  obj,
  t,
  w,
  covs,
  hist = NULL,
  type = "unconditional",
  CI = "MC",
  sig = 0.05,
  MC_B = 100,
  numint_nodes = NULL,
  par_sample = NULL
)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object with fit results.}

\item{t}{non-negative numeric Prediction point (start of predition interval).}

\item{w}{positive numeric Prediction horizon (marks the end of prediction
interval).}

\item{covs}{list with covariate values:
\itemize{
\item \verb{$terminal} = numeric vector Covariates for terminal event model.
\item \verb{$recurrent} = numeric vector Covariates for recurrent events model.
}}

\item{hist}{numeric vector with recurrent event history. Default is \code{c()}
(i.e., \code{NULL}), which corresponds to zero events before \code{t}.}

\item{type}{Whether to predict given recurrent event history or not. Values:
\code{"unconditional"} (default), to predict just conditioning on covariates, or
\code{"conditional"} to predict given recurrent event history and covariates.}

\item{CI}{Compute Confidence Interval? \code{MC} (Default: Monte Carlo) or \code{DM}
(Delta-method). Any other value does not compute the Confidence Interval.}

\item{sig}{Significance of the confidence interval: An interval of \code{1 - sig}
confidence is computed. Default: \code{0.05}.}

\item{MC_B}{Number of Monte Carlo draws. Default is \code{100}.}

\item{numint_nodes}{Number of numerical integration nodes. Default is the
same as for estimation.}

\item{par_sample}{A matrix with a draw from the asymptotic distribution of
parameters (for Monte Carlo Confidence Intervals). Can be generated by
\code{\link[=asympt_draw.SharedModel]{asympt_draw.SharedModel()}}.}
}
\value{
numeric Predicted risk of terminal event. If \code{CI = TRUE}, the
returned values has attributes \code{"CI_lower"} and \code{"CI_upper"}.
}
\description{
This method predicts the risk of terminal event in the interval \eqn{(t, t+w)}
given covariates and recurrent event history.
}
\details{
If \code{CI='MC'}, the Confidence Interval is computed my Monte Carlo. The steps
are: (i) draw a sample from the asymptoticdistribution of the parameter,
(ii) compute the risk of the ternimal event with each drawn parameter, and
(iii) take the \code{sig/2} and \code{1-sig/2} percentiles of those values.

If \code{CI='DM'}, the Confidence Interval is computed my the Delta-method. That is,
the derivative of the expression for the predicted risk (numerically
approximating the integrals) is computed. Then, the variance of the prediction
is estimated multiplying the derivative to the asymptotic variance.

The default number of numerical integration nodes (the same as for estimation
of the model) may be to large if \code{CI='MC'} or if the function must be applied
to a large number of times \code{t} or horizons \code{w}. The user may alter this by
setting \code{numint_nodes}.

If the function is called for several times \code{t} or horizons \code{w}, the user can
provide a sample \code{par_sample} from the asymptotic distribution of the
parameters. This allows fixing the sample acroos all calls. The sample may be
constructed by bootstrapping or by a normal approximation, using
\code{\link[=asympt_draw.SharedModel]{asympt_draw.SharedModel()}}.
}
