% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{predict_plot}
\alias{predict_plot}
\title{Plot with terminal event risk predictions}
\usage{
predict_plot(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=predict_plot.SharedModel]{predict_plot.SharedModel()}} for more information.
}
