% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{print.SharedModel}
\alias{print.SharedModel}
\title{Print fit results}
\usage{
\method{print}{SharedModel}(
  obj,
  cols = c("est", "hr", "pvals"),
  col_names = NULL,
  digits = 3,
  BHHH = TRUE
)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object with fit results.}

\item{cols}{character vector with values to display. A subset of
\code{c('est', 'est_SE', 'est_CI', 'hr', 'hr_SE', 'hr_CI', 'pvals')}.}

\item{col_names}{Names of the columns.}

\item{digits}{Digits to round results.}

\item{BHHH}{\code{TRUE} to use Score based variance for computations. \code{FALSE} to
use the variance based on the optimization routine Hessian.}
}
\value{
Silent. Prints information in console.
}
\description{
Print fit results
}
