% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{rec_boxplot}
\alias{rec_boxplot}
\title{Box plot for number of recurrent events}
\usage{
rec_boxplot(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=rec_boxplot.SharedModel]{rec_boxplot.SharedModel()}} for more information.
}
