% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{rec_boxplot.SharedModel}
\alias{rec_boxplot.SharedModel}
\title{Boxplot with number of recurrent events}
\usage{
\method{rec_boxplot}{SharedModel}(obj, ...)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}

\item{...}{Additional parameters to \code{\link[graphics:boxplot]{graphics::boxplot()}}. Also, specify
\code{horizontal=TRUE} for a horizontal plot.}
}
\value{
Silent. Creates plot.
}
\description{
Boxplot with number of recurrent events
}
