% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{scores}
\alias{scores}
\title{Scores of the model}
\usage{
scores(obj)
}
\description{
This is an S3 generic. See \code{\link[=scores.SharedModel]{scores.SharedModel()}} for more information.
}
