% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{scores.SharedModel}
\alias{scores.SharedModel}
\title{Scores of the model}
\usage{
\method{scores}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
numeric matrix with individuals in rows and scores wrt each
parameter in columns.
}
\description{
Scores of the model
}
