% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_int.R
\name{select.intsample}
\alias{select.intsample}
\title{Sample and weights for numerical integral}
\usage{
select.intsample(pdfinfo, intcontrol)
}
\arguments{
\item{intcontrol}{}
}
\description{
Sample and weights for numerical integral
}
