% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{SharedModel}
\alias{SharedModel}
\title{Build a SharedModel object}
\usage{
SharedModel(
  par,
  varmat,
  par_pos,
  var_names,
  OBSL,
  logLikelihood,
  function_shapes,
  rec_timescale,
  Calls,
  optim_control
)
}
\arguments{
\item{par}{Parameter of the model}

\item{varmat}{Asymptotic covariance matrix.}

\item{par_pos}{list with indexes of subparameters:
\itemize{
\item \verb{$alpha}= frailty parameter.
\item \verb{$sig} = parameters of the frailty distribution.
\item \verb{$a_d} = parameters of the baseline hazard for terminal event.
\item \verb{$a_r} = parameters of the baseline hazard for recurrent events.
\item \verb{$beta_d} = covariate parameters in the terminal event model.
\item \verb{$beta_r} = covariate parameters in the recurrent events model.
\item \verb{$pw_C} = parameters of the piecewise renewal hazard (Poisson constants).
}}

\item{var_names}{list with covariate variable names:
\itemize{
\item \verb{$terminal} = for terminal event model.
\item \verb{$recurrent} = for recurrent events model.
}}

\item{OBSL}{List with observation unit information. See \code{\link[=individual_loglikelihood]{individual_loglikelihood()}}.}

\item{logLikelihood}{Value of the log-likelihood at optimum.}

\item{function_shapes}{list with function shapes for the model:
\itemize{
\item \verb{$hazard_d} = model for baseline hazard of the terminal event.
\item \verb{$hazard_r} = model for baseline hazard of recurrent events.
\item \verb{$surv_d} = model for baseline survival function of the terminal event.
\item \verb{$surv_r} = model for baseline survival function of the recurrent events.
\item \verb{$pdf} = model for the distribution of the frailty variable.
\item \verb{$rec_piecewise_ts} = thresholds for piecewise renewal hazard.
}}

\item{rec_timescale}{Model specification for the hazard of the recurrent event.}

\item{Calls}{list with calls with formulae:
\itemize{
\item \verb{$terminal} = for terminal event model.
\item \verb{$recurrent} = for recurrent events model.
}}

\item{optim_control}{list with information for optimization
\itemize{
\item \verb{$gradtol} = Tolerance for the gradient, see \code{\link[stats:nlm]{stats::nlm()}}.
\item \verb{$steptol} = Tolerance for the steps, see \code{\link[stats:nlm]{stats::nlm()}}.
\item \verb{$iterlim} = Iteration limit, see \code{\link[stats:nlm]{stats::nlm()}}.
\item \verb{$iter} = Number of iterations until convergence.
\item \verb{$nlm_code} = Optimization exit status, see \code{\link[stats:nlm]{stats::nlm()}}.
\item \verb{$nlm_hessian} = Hessian returned by \code{\link[stats:nlm]{stats::nlm()}}.
\item \verb{$nlm_warnings} = Number of warnings during optimization. Mostly due to
expressions evaluating to infinity.
\item \verb{$int_mode} = Numerical integration mode. See \code{\link[=shared_frailty_fit]{shared_frailty_fit()}}.
\item \verb{$MC_N} = Number of points for Monte Carlo integration.
\item \verb{$GC_nodes} = Number of points for Gaussian Quadrature integration.
\item \verb{$is_control} = Importance Sampling control parameters. See \code{\link[=shared_frailty_fit]{shared_frailty_fit()}}.
\item \verb{$anal.grad} = If analytic gradient was used.
\item \verb{$BHHH.hessian} = If BHHH Hessian was used. See \code{\link[=shared_frailty_fit]{shared_frailty_fit()}}.
\item \verb{$posit_cons} = Parameters with positivity constrains (transformations).
\item \verb{$par_scale} = Scale of each parameter.
\item \verb{$MC_samp} (Only during computarions) = Monte Carlo sample for numerical integration.
}}
}
\value{
SharedModel object.
}
\description{
SharedModel S3 objects contain fit information and information regarding
control of the optimization process. See Details for a list of S3 methods.
}
\details{
List of methods:
\itemize{
\item \code{\link[=asymptvar.SharedModel]{asymptvar.SharedModel()}}
\item \code{\link[=asympt_draw.SharedModel]{asympt_draw.SharedModel()}}
\item \code{\link[=base_hazard.SharedModel]{base_hazard.SharedModel()}}
\item \code{\link[=base_survival.SharedModel]{base_survival.SharedModel()}}
\item \code{\link[=BHHH_hessian.SharedModel]{BHHH_hessian.SharedModel()}}
\item \code{\link[=check_intsample.SharedModel]{check_intsample.SharedModel()}}
\item \code{\link[=dist_relevance_test.SharedModel]{dist_relevance_test.SharedModel()}}
\item \code{\link[=get_rec_number.SharedModel]{get_rec_number.SharedModel()}}
\item \code{\link[=hazard_given_history.SharedModel]{hazard_given_history.SharedModel()}}
\item \code{\link[=logLik.SharedModel]{logLik.SharedModel()}}
\item \code{\link[=max_timespan.SharedModel]{max_timespan.SharedModel()}}
\item \code{\link[=optim_details.SharedModel]{optim_details.SharedModel()}}
\item \code{\link[=param_coef.SharedModel]{param_coef.SharedModel()}}
\item \code{\link[=param_frailty.SharedModel]{param_frailty.SharedModel()}}
\item \code{\link[=param_hazard.SharedModel]{param_hazard.SharedModel()}}
\item \code{\link[=param_tables.SharedModel]{param_tables.SharedModel()}}
\item \code{\link[=predict.SharedModel]{predict.SharedModel()}}
\item \code{\link[=predict_plot.SharedModel]{predict_plot.SharedModel()}}
\item \code{\link[=print.SharedModel]{print.SharedModel()}}
\item \code{\link[=rec_boxplot.SharedModel]{rec_boxplot.SharedModel()}}
\item \code{\link[=scores.SharedModel]{scores.SharedModel()}}
\item \code{\link[=summary.SharedModel]{summary.SharedModel()}}
\item \code{\link[=surv_given_history.SharedModel]{surv_given_history.SharedModel()}}
\item \code{\link[=toLatex.SharedModel]{toLatex.SharedModel()}}
\item \code{\link[=Wald_test.SharedModel]{Wald_test.SharedModel()}}
}
}
