% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{summary.SharedModel}
\alias{summary.SharedModel}
\title{Descriptive summary of the sample}
\usage{
\method{summary}{SharedModel}(obj)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}
}
\value{
Silent. Prints information in console.
}
\description{
Descriptive summary of the sample
}
