% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{surv_given_history}
\alias{surv_given_history}
\title{Survival function given recurrent event history}
\usage{
surv_given_history(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=surv_given_history.SharedModel]{surv_given_history.SharedModel()}} for more information.
}
