% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{surv_given_history.SharedModel}
\alias{surv_given_history.SharedModel}
\title{Survival function given history}
\usage{
\method{surv_given_history}{SharedModel}(obj, t, Tjs = NULL, gradient = FALSE)
}
\arguments{
\item{obj}{\code{\link[=SharedModel]{SharedModel()}} object}

\item{t}{time to evaluate the survival function}

\item{Tjs}{numeric vector with recurrent event times prior to \code{t}.}

\item{gradient}{\code{TRUE} to compute the gradient (default is \code{FALSE}).}
}
\value{
numeric value of the survival function given history at \code{t}. If
\code{gradient=TRUE}, the returned value has an attribute \code{"gradient"} with its
derivative.
}
\description{
Finds the survival function of the recurrent event at \code{t}, given the history
of recurrent events prior to \code{t}.
}
