% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_methods.R
\name{times.Surv}
\alias{times.Surv}
\title{Get observation time}
\usage{
\method{times}{Surv}(obj)
}
\arguments{
\item{obj}{\code{\link[survival:Surv]{survival::Surv()}} object}
}
\value{
numeric time value in object
}
\description{
Find the observation time from \code{\link[survival:Surv]{survival::Surv()}} object.
}
