% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_generics.R
\name{uncensored}
\alias{uncensored}
\title{Get censoring information}
\usage{
uncensored(obj)
}
\description{
This is an S3 generic. See \code{\link[=uncensored.Surv]{uncensored.Surv()}} for more information.
}
