% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_methods.R
\name{uncensored.Surv}
\alias{uncensored.Surv}
\title{Get censoring information}
\usage{
\method{uncensored}{Surv}(obj)
}
\arguments{
\item{obj}{\code{\link[survival:Surv]{survival::Surv()}} object}
}
\value{
boolean \code{TRUE} if observation is not censored
}
\description{
Find censoring information from \code{\link[survival:Surv]{survival::Surv()}} object.
}
