% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fit.R
\name{valid_posit_constraint}
\alias{valid_posit_constraint}
\title{Valid positivity constraint}
\usage{
valid_posit_constraint(par_pos, model_funs)
}
\arguments{
\item{par_pos}{list with indexes of subparameters, see \code{\link[=SharedModel]{SharedModel()}}.}

\item{model_funs}{list with function shapes for the model, see \code{\link[=SharedModel]{SharedModel()}}.}
}
\value{
numeric vector with constraint parameter positions
}
\description{
Get the positivity constraints for each function in the model.
}
