% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{Wald_test}
\alias{Wald_test}
\title{Wald tests for the parameters of the model}
\usage{
Wald_test(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=Wald_test.SharedModel]{Wald_test.SharedModel()}} for more information.
}
