% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_methods.R
\name{Wald_test.SharedModel}
\alias{Wald_test.SharedModel}
\title{Wald tests for the parameters of the model}
\usage{
\method{Wald_test}{SharedModel}(obj, termpos, par0 = NULL, BHHH = TRUE)
}
\arguments{
\item{obj}{obj \code{\link[=SharedModel]{SharedModel()}} object}

\item{termpos}{numeric vector with positions of the terms to test. Use
\code{obj$par_pos}. See \code{\link[=SharedModel]{SharedModel()}}.}

\item{par0}{Null hypothesis to test.}

\item{BHHH}{\code{TRUE} to use Score based variance for computations. \code{FALSE} to
use the variance based on the optimization routine Hessian.}
}
\value{
An object of class \link{wald.test}. See \code{\link[aod:wald.test]{aod::wald.test()}}.
}
\description{
This function is a wrapper around \code{\link[aod:wald.test]{aod::wald.test()}}.
}
