% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedModel_generics.R
\name{param_tables}
\alias{param_tables}
\title{Parameter tables for printing and latex output}
\usage{
param_tables(obj, ...)
}
\description{
This is an S3 generic. See \code{\link[=param_tables.SharedModel]{param_tables.SharedModel()}} for more information.
}
